requirements.txt
pandas

pymongo
python-dotenv
langchain
langchain-openai
langgraph
streamlit

langchain_community
PyPDF2
